Moves proAdhoc extra functions to a separate file
